Move info to src
