+++ pygmt/src/__init__.py
@@ -6,6 +6,7 @@ from pygmt.src.blockmedian import blockmedian
 from pygmt.src.grdcut import grdcut
 from pygmt.src.grdfilter import grdfilter
 from pygmt.src.grdtrack import grdtrack
+from pygmt.src.info import info
 from pygmt.src.makecpt import makecpt
 from pygmt.src.meca import meca
 from pygmt.src.surface import surface
